CAS-1341: Offer the option to provide a new HTTP client implementation
First commit
