+++ cas-server-core/src/test/java/org/jasig/cas/authentication/handler/support/HttpBasedServiceCredentialsAuthenticationHandlerTests.java
@@ -21,7 +21,7 @@ package org.jasig.cas.authentication.handler.support;
 import javax.security.auth.login.FailedLoginException;
 
 import org.jasig.cas.TestUtils;
-import org.jasig.cas.util.HttpClient;
+import org.jasig.cas.util.SimpleHttpClient;
 
 import org.junit.Before;
 import org.junit.Test;
@@ -39,7 +39,7 @@ public final class HttpBasedServiceCredentialsAuthenticationHandlerTests {
     @Before
     public void setUp() throws Exception {
         this.authenticationHandler = new HttpBasedServiceCredentialsAuthenticationHandler();
-        this.authenticationHandler.setHttpClient(new HttpClient());
+        this.authenticationHandler.setHttpClient(new SimpleHttpClient());
     }
 
     @Test
@@ -70,7 +70,7 @@ public final class HttpBasedServiceCredentialsAuthenticationHandlerTests {
 
     @Test
     public void testAcceptsNonHttpsCredentials() throws Exception {
-        this.authenticationHandler.setHttpClient(new HttpClient());
+        this.authenticationHandler.setHttpClient(new SimpleHttpClient());
         this.authenticationHandler.setRequireSecure(false);
         assertNotNull(this.authenticationHandler.authenticate(
                 TestUtils.getHttpBasedServiceCredentials("http://www.jasig.org")));
@@ -84,7 +84,7 @@ public final class HttpBasedServiceCredentialsAuthenticationHandlerTests {
 
     @Test(expected = FailedLoginException.class)
     public void testNoAcceptableStatusCodeButOneSet() throws Exception {
-        final HttpClient httpClient = new HttpClient();
+        final SimpleHttpClient httpClient = new SimpleHttpClient();
         httpClient.setAcceptableCodes(new int[] {900});
         this.authenticationHandler.setHttpClient(httpClient);
         this.authenticationHandler.authenticate(TestUtils.getHttpBasedServiceCredentials("https://www.ja-sig.org"));
