+++ cas-server-core/src/test/resources/core-context.xml
@@ -49,7 +49,7 @@
             value-ref="uniqueTicketIdGenerator"/>
     </util:map>
 
-    <bean id="httpClient" class="org.jasig.cas.util.HttpClient"
+    <bean id="httpClient" class="org.jasig.cas.util.SimpleHttpClient"
           p:readTimeout="5000"
           p:connectionTimeout="5000" />
