+++ cas-server-core/src/test/resources/mfa-test-context.xml
@@ -83,7 +83,7 @@
         </property>
     </bean>
 
-    <bean id="httpClient" class="org.jasig.cas.util.HttpClient"
+    <bean id="httpClient" class="org.jasig.cas.util.SimpleHttpClient"
           p:readTimeout="5000"
           p:connectionTimeout="5000" />
