+++ cas-server-integration-memcached/src/main/java/org/jasig/cas/ticket/registry/support/kryo/serial/SamlServiceSerializer.java
@@ -25,6 +25,7 @@ import com.esotericsoftware.kryo.Kryo;
 import org.jasig.cas.support.saml.authentication.principal.SamlService;
 import org.jasig.cas.ticket.registry.support.kryo.FieldHelper;
 import org.jasig.cas.util.HttpClient;
+import org.jasig.cas.util.SimpleHttpClient;
 
 /**
  * Serializer for {@link SamlService} class.
@@ -61,7 +62,7 @@ public final class SamlServiceSerializer extends AbstractWebApplicationServiceSe
 
         final String requestId = kryo.readObject(buffer, String.class);
         try {
-            return (SamlService) CONSTRUCTOR.newInstance(id, originalUrl, artifactId, new HttpClient(), requestId);
+            return (SamlService) CONSTRUCTOR.newInstance(id, originalUrl, artifactId, new SimpleHttpClient(), requestId);
         } catch (final Exception e) {
             throw new IllegalStateException("Error creating SamlService", e);
         }
